+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 -   [[#37]](https://github.com/Marcin-Migdal/m-component-library/issues/37) **[PATCH]** Alert close only when overlay click started on overlay
 -   [[#36]](https://github.com/Marcin-Migdal/m-component-library/issues/36) **[PATCH]** Add hslToHex helper function
 -   [[#34]](https://github.com/Marcin-Migdal/m-component-library/issues/34) **[PATCH]** Fix clear icon in dropdown wrong color
+-   [[#35]](https://github.com/Marcin-Migdal/m-component-library/issues/35) **[PATCH]** Add noBottomMargin props to all inputs that on default have bottom margin
 
 ### 28.09-2024 - Release 1.0.3